requirements.txt
streamlit==1.29.0
openai-whisper==20231024
google-generativeai==0.2.2
python-dotenv==1.0.0
numpy>=1.24.0
